+++ reducer-todo/src/App.js
@@ -1,31 +1,16 @@
- import React from 'react'
- import TodoForm from "./components/TodoForm"
- import './App.css';
- import Todo from "./components/todo"
- import {todos} from "./components/reducer/reducer"
+import React from "react";
+import TodoForm from "./components/TodoForm";
+import "./App.css";
 
 function App() {
-  const todos= [
-   {
-    todo: "learn about reducers",
-    complete:false,
-    id: 1
-  }]
-
-
-
   
-  return (
-    
-      <div className="App">
-        <div className="header">
-           <h1>LIST OF TODOS</h1>
-           <TodoForm todos={todos}/>
-           
-         </div>
-               
-      
 
+  return (
+    <div className="App">
+      <div className="header">
+        <h1>LIST OF TODOS</h1>
+        <TodoForm/>
+      </div>
     </div>
   );
 }
