+++ reducer-todo/src/components/todo.js
@@ -2,22 +2,22 @@ import React  from 'react'
 import { ACTIONS }from "./action"
 
 
-export default function Todo(props) {
-const {dispatch,todos} =props
+export default function Todo({todo,dispatch}) {
 
-    console.log(props)
+
+    
     return (
         <>
         <div>
-      <span  className={`item ${props.todos.complete ? ' complete' : ''}`} >   
-            {todos.name}
+      <span  className={`item ${todo.complete ? ' complete' : ''}`} >   
+            {todo.name}
         </span> 
-            <button onClick={()=>dispatch({type:ACTIONS.TOGGLE_TODO, payload:{id: todos.id} })}>
+            <button onClick={()=>dispatch({type:ACTIONS.TOGGLE_TODO, payload:{id: todo.id} })}>
                 TOGGLE TODO
             </button>
-            ADD TODO
-            <button onClick={()=> dispatch({type: ACTIONS.ADD_TODO, payload: {id:todos.id}})}> ADD TODO</button>
-            <button onClick={()=> dispatch({type: ACTIONS.DELETE_TODO, payload: {id:todos.id}})}>DELETE</button>
+        
+            
+            <button onClick={()=> dispatch({type: ACTIONS.DELETE_TODO, payload: {id:todo.id}})}>DELETE TODO</button>
         </div>
         </>
     )
